+++ framework/src/main/java/org/checkerframework/framework/util/typeinference8/types/InferenceMethodType.java
@@ -3,10 +3,14 @@ package org.checkerframework.framework.util.typeinference8.types;
 import com.sun.source.tree.ExpressionTree;
 import com.sun.source.tree.MethodInvocationTree;
 import com.sun.source.tree.NewClassTree;
+import java.util.ArrayList;
+import java.util.List;
 import javax.lang.model.element.Element;
+import javax.lang.model.type.ArrayType;
 import javax.lang.model.type.ExecutableType;
 import javax.lang.model.type.TypeMirror;
 import org.checkerframework.framework.type.AnnotatedTypeMirror;
+import org.checkerframework.framework.type.AnnotatedTypeMirror.AnnotatedArrayType;
 import org.checkerframework.framework.type.AnnotatedTypeMirror.AnnotatedExecutableType;
 import org.checkerframework.framework.util.typeinference8.util.Java8InferenceContext;
 import org.checkerframework.framework.util.typeinference8.util.Theta;
@@ -19,6 +23,9 @@ import org.checkerframework.javacutil.TreeUtils;
  * AnnotatedExecutableType}
  */
 public class InferenceMethodType extends InferenceExecutableType {
+
+  /** The {@code NewClassTree} or {@code MethodInvocationTree} whose type this is. */
+  private final ExpressionTree invocation;
 
   /**
    * Creates an invocation type for a method or constructor invocation.
@@ -34,6 +41,7 @@ public class InferenceMethodType extends InferenceExecutableType {
       ExpressionTree invocation,
       Java8InferenceContext context) {
     super(annotatedExecutableType, methodType, invocation, context);
+    this.invocation = invocation;
     assert invocation instanceof MethodInvocationTree || invocation instanceof NewClassTree;
   }
 
@@ -64,5 +72,27 @@ public class InferenceMethodType extends InferenceExecutableType {
       return new ProperType(annotatedReturnType, returnType, context);
     }
     return InferenceType.create(annotatedReturnType, returnType, map, context);
+  }
+
+  @Override
+  public List<AbstractType> getParameterTypes(Theta map, int size) {
+    List<AnnotatedTypeMirror> params = new ArrayList<>(annotatedExecutableType.getParameterTypes());
+    List<TypeMirror> paramsJava = new ArrayList<>(methodType.getParameterTypes());
+
+    if (TreeUtils.isVarargsCall(invocation)) {
+      AnnotatedArrayType vararg = (AnnotatedArrayType) params.remove(params.size() - 1);
+      for (int i = params.size(); i < size; i++) {
+        params.add(vararg.getComponentType());
+      }
+    }
+
+    if (TreeUtils.isVarargsCall(invocation)) {
+      ArrayType vararg = (ArrayType) paramsJava.remove(paramsJava.size() - 1);
+      for (int i = paramsJava.size(); i < size; i++) {
+        paramsJava.add(vararg.getComponentType());
+      }
+    }
+
+    return InferenceType.create(params, paramsJava, map, qualifierVars, context);
   }
 }
